+++ src/main/java/com/daylily/domain/github/service/PullRequestHandlerServiceImpl.java
@@ -1,8 +1,9 @@
-package com.daylily.domain.webhook.service;
+package com.daylily.domain.github.service;
 
 import com.daylily.domain.docker.client.DockerGrpcClient;
 import com.daylily.domain.github.action_type.PullRequestActionType;
-import com.daylily.domain.webhook.util.PayloadParser;
+import com.daylily.domain.github.util.ActionTypeChecker;
+import com.daylily.domain.github.util.PayloadParser;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.kohsuke.github.GHEventPayload;
@@ -25,11 +26,12 @@ public class PullRequestHandlerServiceImpl implements PullRequestHandlerService 
     @Override
     public void handlePullRequestEvent(String eventType, String rawPayload) {
         // 1. PR Payload 읽기
-        GHEventPayload.PullRequest payload = payloadParser.parsePullRequestPayload(rawPayload);
+        GHEventPayload.PullRequest payload = payloadParser.parsePullRequest(rawPayload);
 
         // 2. PR Action Type에 따라 처리
         GHPullRequest pullRequest;
-        switch (PullRequestActionType.fromString(payload.getAction())) {
+        PullRequestActionType actionType = ActionTypeChecker.fromPullRequestActionString(payload.getAction());
+        switch (actionType) {
             // TODO: PR 컨테이너는 어떤 경우에 만들어야할지 논의
             // 새 PR 또는 PR에 새 커밋이 푸시된 경우에만 이미지를 빌드하고 컨테이너를 실행합니다.
             case EDITED: // TODO: PR의 Base 브랜치가 변경된 경우에도 이미지를 빌드해야할지 논의
