+++ src/main/java/com/daylily/domain/github/util/WebhookDispatcher.java
@@ -1,4 +1,4 @@
-package com.daylily.domain.webhook.util;
+package com.daylily.domain.github.util;
 
 // TODO: 메시징 큐를 흉내낸 인터페이스이므로, 추후 메시징 큐 도입 시 코드 수정 예정
 public interface WebhookDispatcher {
